Move text-util module to lib
