+++ lib/src/index.ts
@@ -3,3 +3,4 @@ export * from './settings';
 export * from './coqview';
 export * from './AnnotatedText';
 export * from './nodejs-async';
+export * from './text-util';
