+++ server/src/parsing/SentenceSemantics.ts
@@ -1,7 +1,7 @@
 import * as vscode from 'vscode-languageserver';
 import * as server from '../server';
 import * as parser from './coq-parser';
-import * as textUtil from './../util/text-util';
+import * as textUtil from '@lib/text-util';
 import * as path from 'path';
 import * as fs from 'fs';
