+++ server/src/stm/State.ts
@@ -2,7 +2,7 @@ import {Position, Range, DiagnosticSeverity} from 'vscode-languageserver';
 import * as vscode from 'vscode-languageserver';
 import * as coqProto from './../coqtop/coq-proto';
 import * as parser from './../parsing/coq-parser';
-import * as textUtil from './../util/text-util';
+import * as textUtil from '@lib/text-util';
 import { ProofView } from '@lib/protocol';
 import {AnnotatedText} from '../util/AnnotatedText';
 import {ProofViewReference, GoalsCache} from './GoalsCache'
